Add new pitt notifications
