+++ src/base_app/templates/new_pitt_notice_title.txt
@@ -1,0 +1,1 @@
+New pitt from {{pitt.user.username}}
